+++ benchmarks/verus/numpy_triple_part_2/files/polynomial_chebyshev_chebgauss.rs
@@ -8,16 +8,16 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn chebgauss(n: usize) -> (result: (Vec<f32>, Vec<f32>))
+fn chebgauss(n: u8) -> (result: (Vec<f32>, Vec<f32>))
     requires n > 0,
     ensures
-        result.0.len() == n,
-        result.1.len() == n,
+        result.0.len() == n as usize,
+        result.1.len() == n as usize,
         /* All weights are equal (uniform weights) */
-        forall|i: int, j: int| 0 <= i < n && 0 <= j < n ==> 
+        forall|i: int, j: int| 0 <= i < n as int && 0 <= j < n as int ==> 
             #[trigger] result.1[i] == #[trigger] result.1[j],
         /* Nodes are distinct */
-        forall|i: int, j: int| 0 <= i < n && 0 <= j < n && i != j ==> 
+        forall|i: int, j: int| 0 <= i < n as int && 0 <= j < n as int && i != j ==> 
             #[trigger] result.0[i] != #[trigger] result.0[j],
 // </vc-spec>
 // <vc-code>
