+++ benchmarks/verus/numpy_triple_part_2/files/statistics_histogram_bin_edges.rs
@@ -8,7 +8,7 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn histogram_bin_edges(data: Vec<i32>, num_bins: usize) -> (edges: Vec<i32>)
+fn histogram_bin_edges(data: Vec<i8>, num_bins: usize) -> (edges: Vec<i8>)
     requires 
         data.len() > 0,
         num_bins > 0,
